pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rna_map"
version = "0.4.1"
description = "rapid analysis of RNA mutational profiling (MaP) experiments."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
  { name = "Joe Yesselman", email = "jyesselm@unl.edu" }
]
maintainers = [
  { name = "Joe Yesselman", email = "jyesselm@unl.edu" }
]
keywords = ["rna_map", "RNA", "mutational profiling", "MaP", "bioinformatics"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "click>=8",
  "cloup>=2.0",
  "colorlog>=6.0",
  "jsonschema>=4.0",
  "matplotlib>=3.5",
  "numpy>=1.21",
  "pandas>=1.5",
  "plotly>=5.0",
  "pyyaml>=6.0",
  "tabulate>=0.9",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "pytest-cov>=4.0",
  "pytest-mock>=3.0",
  "ruff>=0.14",
  "mypy>=1.0",
  "interrogate>=1.7",
  "radon>=6.0",
  "pylint>=3.0",
]
dask = [
  "dask>=2023.0.0",
  "distributed>=2023.0.0",
  "dask-jobqueue>=0.8.0",
]
dev = [
  "tox>=4.0",
]

[project.urls]
Homepage = "https://github.com/YesselmanLab/rna_map"
Repository = "https://github.com/YesselmanLab/rna_map"
Documentation = "https://github.com/YesselmanLab/rna_map#readme"
Issues = "https://github.com/YesselmanLab/rna_map/issues"

[project.scripts]
rna-map = "rna_map.cli:cli"
rna-map-pipeline = "rna_map.cli_pipeline:pipeline"

[tool.setuptools.packages.find]
include = ["rna_map*"]

[tool.setuptools.package-data]
rna_map = ["py.typed", "resources/*", "resources/**/*"]

# Ruff configuration (replaces black, isort, flake8)
[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    "build",
    "dist",
    "*.egg-info",
    ".tox",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "W",   # pycodestyle warnings
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "D",   # pydocstyle (docstrings)
    "SIM", # flake8-simplify
    "C90", # mccabe complexity
]
extend-ignore = [
    "E203",  # whitespace before ':' (conflicts with formatter)
    "E501",  # Line too long (handled by formatter)
    "D100",  # Missing docstring in public module
    "D104",  # Missing docstring in public package
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["rna_map"]
force-single-line = false
split-on-trailing-comma = true
combine-as-imports = true
force-wrap-aliases = false
force-sort-within-sections = true
order-by-type = true

[tool.ruff.lint.pydocstyle]
convention = "google"
ignore-decorators = ["typing.overload", "functools.lru_cache"]
property-decorators = ["property", "cached_property"]

[tool.ruff.lint.per-file-ignores]
"test/**" = [
    "D103",  # Missing docstring in public function
    "D101",  # Missing docstring in public class
    "D102",  # Missing docstring in public method
    "D107",  # Missing docstring in __init__
    "D205",  # 1 blank line required between summary line and description
]

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# MyPy configuration (less strict than before)
[tool.mypy]
python_version = "3.10"
packages = ["rna_map"]
strict = false
warn_unused_configs = true
# Allow functions without type annotations (less aggressive)
disallow_untyped_defs = false
# Don't require complete type annotations
disallow_incomplete_defs = false
# Allow implicit Optional types (e.g., str | None can be just str)
no_implicit_optional = false
# Don't check bodies of untyped functions
check_untyped_defs = false
# Don't warn about returning Any
warn_return_any = false
# Don't warn about unused type: ignore comments
warn_unused_ignores = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "plotly.*",
    "matplotlib.*",
    "pandas.*",
    "numpy.*",
    "click.*",
    "cloup.*",
    "colorlog.*",
    "tabulate.*",
    "yaml.*",
    "jsonschema.*",
]
ignore_missing_imports = true

# Interrogate configuration (docstring coverage)
[tool.interrogate]
fail-under = 90
ignore-init-module = false
ignore-init-method = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = true
ignore-property-decorators = false
ignore-protected = false
ignore-module = false
ignore-nested-functions = false
ignore-nested-classes = false
exclude = [
    "venv",
    ".venv",
    ".git",
    "build",
    "dist",
    "*.egg-info",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    "__pycache__",
]
ignore-regex = [
    "^test_",
    "^_",
]
verbose = 2
quiet = false
whitelist-regex = []

# Radon configuration (complexity checking)
[tool.radon]
# Cyclomatic complexity thresholds
cc_min = "A"  # Minimum grade (A=low complexity, F=very high)
cc_max = "B"  # Maximum acceptable grade

# Maintainability index thresholds
mi_min = 20  # Minimum maintainability index (0-100, higher is better)
mi_avg = 40  # Average maintainability index

# Exclude patterns
exclude = [
    "venv",
    ".venv",
    ".git",
    "build",
    "dist",
    "*.egg-info",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    "__pycache__",
    "test",
]

# Pytest configuration
[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --strict-config --cov=rna_map --cov-report=term-missing --cov-report=html"
testpaths = ["test"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "quick: marks tests as quick unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["rna_map"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/build/*",
    "*/dist/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
fail_under = 90
